filter pipe, modals, snackbar, soyAdmin(), layout

+++ Web/fire/src/app/gestores/gestores.component.ts
@@ -1,10 +1,18 @@
+import { DialogComponent } from '../dialog/dialog.component';
 import { Gestor } from '../gestor';
 import { Component, OnInit } from '@angular/core';
-import {GestoresService} from '../gestores.service';
+import { GestoresService } from '../gestores.service';
+import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatTableDataSource } from '@angular/material/table';
 import { Data } from '@angular/router';
 import { User } from 'firebase';
-import { Observable } from 'rxjs';
+import { Observable } from 'rxjs/';
+import { filter } from 'rxjs/operators';
+
+/**
+ * @title Snack-bar with a custom component
+ */
 
 @Component({
   selector: 'app-gestores',
@@ -19,11 +27,14 @@ export class GestoresComponent implements OnInit {
   name: string;
   email: string;
   password: string;
-
-  constructor(private gestoresService: GestoresService) {
+  filter: string;
+  
+  constructor(private gestoresService: GestoresService, public snackBar: MatSnackBar, public dialog: MatDialog) {
     this.nuevoGestor = false;
     this.modificarGestor = false;
     this.getList();
+    this.name = '';
+    this.email = '';
    }
 
   newGestor() {
@@ -34,36 +45,57 @@ export class GestoresComponent implements OnInit {
     }
   }
 
-  modifyGestor(name: string) {
+  modifyGestor(name: string, email: string) {
     if (this.modificarGestor) {
       this.modificarGestor = false;
       this.name = '';
     } else {
       this.modificarGestor = true;
+      this.email = email;
       this.name = name;
     }
   }
 
   getList() {
     this.gestoresList = this.gestoresService.getList();
-    console.log("GestoresList"); 
-    console.log(this.gestoresList);
   }
 
   addGestor() {
-    this.gestoresService.addGestor(this.name, this.email, this.password);
+    this.gestoresService.addGestor(this.name, this.email, '124wqerQWE');
+    this.openSnackBar(this.name + ' añadido con éxtio!', '');
     this.name = '';
     this.email = '';
     this.password = '';
+    this.newGestor();
     }
-  
+
   deleteGestor(uid: string) {
-    this.gestoresService.deleteGestor(uid);
+    const dialog = this.dialog.open(DialogComponent, {
+      data: {
+        mensaje: '¿Está seguro de que desea eliminar el gestor?',
+        accion: 'Confirmar'
+      }
+    });
+
+    dialog.afterClosed().subscribe(result => {
+      if (result) {
+        this.gestoresService.deleteGestor(uid);
+        this.openSnackBar('Eliminado con éxito!', 'OK');
+      }
+    });
   }
 
 
   ngOnInit() {
   }
+
+  openSnackBar(message: string, action: string) {
+    this.snackBar.open(message, action, {
+      duration: 2000
+    });
+  }
+
 }
+
